+++ vaadin-time-picker-flow-parent/vaadin-time-picker-flow-integration-tests/src/main/java/com/vaadin/flow/component/timepicker/tests/validation/BasicValidationPage.java
@@ -7,13 +7,13 @@ import com.vaadin.flow.router.Route;
 import com.vaadin.tests.validation.AbstractValidationPage;
 
 @Route("vaadin-time-picker/validation/basic")
-public class TimePickerValidationBasicPage
+public class BasicValidationPage
         extends AbstractValidationPage<TimePicker> {
     public static final String REQUIRED_BUTTON = "required-button";
     public static final String MIN_INPUT = "min-input";
     public static final String MAX_INPUT = "max-input";
 
-    public TimePickerValidationBasicPage() {
+    public BasicValidationPage() {
         super();
 
         add(createButton(REQUIRED_BUTTON, "Enable required", event -> {
