+++ vaadin-time-picker-flow-parent/vaadin-time-picker-flow-integration-tests/src/test/java/com/vaadin/flow/component/timepicker/tests/validation/BinderValidationIT.java
@@ -7,74 +7,62 @@ import com.vaadin.tests.validation.AbstractValidationIT;
 import org.junit.Test;
 import org.openqa.selenium.Keys;
 
-import static com.vaadin.flow.component.timepicker.tests.validation.TimePickerValidationBasicPage.MIN_INPUT;
-import static com.vaadin.flow.component.timepicker.tests.validation.TimePickerValidationBasicPage.MAX_INPUT;
-import static com.vaadin.flow.component.timepicker.tests.validation.TimePickerValidationBasicPage.REQUIRED_BUTTON;
+import static com.vaadin.flow.component.timepicker.tests.validation.BinderValidationPage.MIN_INPUT;
+import static com.vaadin.flow.component.timepicker.tests.validation.BinderValidationPage.MAX_INPUT;
+import static com.vaadin.flow.component.timepicker.tests.validation.BinderValidationPage.EXPECTED_VALUE_INPUT;
+import static com.vaadin.flow.component.timepicker.tests.validation.BinderValidationPage.REQUIRED_ERROR_MESSAGE;
+import static com.vaadin.flow.component.timepicker.tests.validation.BinderValidationPage.UNEXPECTED_VALUE_ERROR_MESSAGE;
 
-@TestPath("vaadin-time-picker/validation/basic")
-public class TimePickerValidationBasicIT
+@TestPath("vaadin-time-picker/validation/binder")
+public class BinderValidationIT
         extends AbstractValidationIT<TimePickerElement> {
     @Test
     public void fieldIsInitiallyValid() {
         assertClientValid();
         assertServerValid();
-    }
-
-    @Test
-    public void detach_attach_preservesInvalidState() {
-        // Make field invalid
-        $("button").id(REQUIRED_BUTTON).click();
-        testField.sendKeys(Keys.TAB);
-
-        detachAndReattachField();
-
-        assertServerInvalid();
-        assertClientInvalid();
-    }
-
-    @Test
-    public void webComponentCanNotModifyInvalidState() {
-        assertWebComponentCanNotModifyInvalidState();
-
-        detachAndReattachField();
-
-        assertWebComponentCanNotModifyInvalidState();
+        assertErrorMessage(null);
     }
 
     @Test
     public void required_triggerInputBlur_assertValidity() {
-        $("button").id(REQUIRED_BUTTON).click();
-
         testField.sendKeys(Keys.TAB);
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
     }
 
     @Test
     public void required_changeInputValue_assertValidity() {
-        $("button").id(REQUIRED_BUTTON).click();
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("10:00", Keys.ENTER);
 
-        testField.selectByText("12:00");
+        testField.selectByText("10:00");
         assertServerValid();
         assertClientValid();
 
         testField.selectByText("");
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
     }
 
     @Test
     public void min_changeInputValue_assertValidity() {
         $("input").id(MIN_INPUT).sendKeys("11:00", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("12:00", Keys.ENTER);
 
+        // Constraint validation fails:
         testField.selectByText("10:00");
         assertClientInvalid();
         assertServerInvalid();
+        assertErrorMessage("");
 
+        // Binder validation fails:
         testField.selectByText("11:00");
-        assertClientValid();
-        assertServerValid();
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
 
+        // Both validations pass:
         testField.selectByText("12:00");
         assertClientValid();
         assertServerValid();
@@ -83,15 +71,21 @@ public class TimePickerValidationBasicIT
     @Test
     public void max_changeInputValue_assertValidity() {
         $("input").id(MAX_INPUT).sendKeys("11:00", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("10:00", Keys.ENTER);
 
+        // Constraint validation fails:
         testField.selectByText("12:00");
         assertClientInvalid();
         assertServerInvalid();
+        assertErrorMessage("");
 
+        // Binder validation fails:
         testField.selectByText("11:00");
-        assertClientValid();
-        assertServerValid();
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
 
+        // Both validations pass:
         testField.selectByText("10:00");
         assertClientValid();
         assertServerValid();
@@ -99,9 +93,12 @@ public class TimePickerValidationBasicIT
 
     @Test
     public void badInput_changeInputValue_assertValidity() {
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("10:00", Keys.ENTER);
+
         testField.selectByText("INVALID");
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage("");
 
         testField.selectByText("10:00");
         assertServerValid();
@@ -110,6 +107,7 @@ public class TimePickerValidationBasicIT
         testField.selectByText("INVALID");
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage("");
     }
 
     protected TimePickerElement getTestField() {
